Revert "退出逻辑优化"
